+++ security/oauth2/src/test/java/testes/RunOAuth2TestServer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package leap.oauth2;
+package testes;
 
 import leap.webunit.server.TWebServer;
