+++ security/oauth2/src/test/java/testes/as/RefreshTokenTest.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package leap.oauth2.as;
+package testes.as;
 
-import leap.oauth2.TokenResponse;
-import leap.oauth2.OAuth2TestBase;
+import testes.TokenResponse;
+import testes.OAuth2TestBase;
 
 import org.junit.Test;
