+++ security/oauth2/src/test/java/testes/rs/AdminControllerTest.java
@@ -13,15 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package leap.oauth2.wac.token;
+package testes.rs;
 
-import leap.lang.expirable.TimeExpirable;
+import testes.OAuth2TestBase;
+import testes.TokenResponse;
 
-public interface WacAccessToken extends TimeExpirable {
+import org.junit.Test;
 
-    String getToken();
+public class AdminControllerTest extends OAuth2TestBase {
+
+    @Test
+    public void testClientOnlyAccessToken() {
+        serverContextPath = "/server";
+        
+        TokenResponse token = obtainAccessTokenByClient("client1", "client1_secret");
+        
+        withAccessToken(forGet("/resapp/book"), token.accessToken).send().assertNotOk();
+        withAccessToken(forGet("/resapp/admin/hello"), token.accessToken).send().assertOk();
+    }
     
-    String getRefreshToken();
-
-    String getUserId();
 }
