+++ security/web/src/main/java/leap/web/security/login/DefaultLoginManager.java
@@ -17,6 +17,8 @@ package leap.web.security.login;
 
 import leap.core.annotation.Inject;
 import leap.lang.intercepting.State;
+import leap.lang.logging.Log;
+import leap.lang.logging.LogFactory;
 import leap.web.Request;
 import leap.web.Response;
 import leap.web.security.DefaultSecurityContextHolder;
@@ -28,6 +30,8 @@ import leap.web.security.authc.AuthenticationManager;
 import leap.web.security.authc.SimpleAuthentication;
 
 public class DefaultLoginManager implements LoginManager {
+
+    private static final Log log = LogFactory.get(DefaultLoginManager.class);
     
     protected @Inject SecurityConfig        config;
     protected @Inject LoginHandler[]       handlers;
@@ -44,10 +48,13 @@ public class DefaultLoginManager implements LoginManager {
         }
         
         if(request.isAjax()) {
+            log.debug("Promote login for ajax request");
             ajaxHandler.promoteLogin(request, response, context);
         }else{
+            log.debug("Promote login");
             viewHandler.promoteLogin(request, response, context);    
         }
+
         return true;
     }
